On 32 bit hardware (iPhone 4S) the milliseconds stuff overflows on 32 bit NSInteger. Use long long for now.

+++ Geocube/MyTools.h
@@ -39,7 +39,7 @@ typedef NS_ENUM(NSInteger, PlaySound) {
 
 + (NSInteger)secondsSinceEpochFromWindows:(NSString *)datetime;
 + (NSInteger)secondsSinceEpochFromISO8601:(NSString *)datetime;
-+ (NSInteger)millisecondsSinceEpoch;
++ (long long)millisecondsSinceEpoch;
 
 + (NSString *)dateTimeString_YYYY_MM_DDThh_mm_ss:(NSInteger)seconds;
 + (NSString *)dateTimeString_YYYY_MM_DDThh_mm_ss;
